Add files via upload
The code takes a sample of text, generates sentence embeddings using a pre-trained model, reduces the dimensionality of those embeddings using UMAP, and clusters the tweets based on their embedding similarity using HDBSCAN. It then assesses the quality of the clusters based on the number of clusters and a probabilistic measure.  The code is structured to allow exploration of different models (though most are commented out in the provided code),  UMAP parameters and HDBSCAN parameters for finding optimal clustering results.