Add files via upload
This code performs text classification on a dataset of tweets, leveraging two distinct datasets: one labeled '9' for positive sentiment and another labeled '0' for negative sentiment. The preprocessing steps include the removal of URLs, mentions, and hashtags, alongside tokenization and stop word elimination, which prepare the tweets for analysis. The text data is then transformed into numerical features using TF-IDF vectorization. To classify the tweets, a Multinomial Naive Bayes classifier is trained, and its performance is assessed through various metrics, including classification reports, ROC AUC, average precision, and precision-recall curves. Furthermore, the code employs Hyperopt for Bayesian optimization to fine-tune the hyperparameters of the Multinomial Naive Bayes classifier with TF-IDF preprocessing, aiming to enhance classification accuracy. In addition, an SGDClassifier is trained and evaluated on the same dataset, utilizing similar performance metrics. The primary objective of this work is to compare the effectiveness of different classification models and hyperparameter tuning techniques in the context of tweet sentiment classification.