Add files via upload
This script automates the hyperparameter optimization process for a Multinomial Naive Bayes text classifier by utilizing Bayesian optimization alongside a designated alpha value for model training. It systematically evaluates the classifier's performance and records the outcomes in a CSV file for further analysis. 

Please note that this code is used for exploratory purposes